refactor: extract HTTP API into crates/api
