+++ crates/api/google/src/commands/brightness_absolute.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-use crate::endpoints::google::device::google_to_internal_brightness;
+use crate::device::google_to_internal_brightness;
 use application::SceneRuntime;
 
 use super::super::request::ExecuteCommandType;
