+++ crates/api/google/src/commands/color_absolute.rs
