+++ crates/api/google/src/commands/color_loop.rs
