+++ crates/api/google/src/commands/stop_effect.rs
