+++ crates/api/src/device.rs
