+++ crates/api/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod device;
+pub mod effects;
