--- src/endpoints/mod.rs
@@ -1,5 +1,0 @@
-pub mod device;
-pub mod effects;
-pub mod google;
-pub mod legacy;
-pub mod oauth;
